Added new pdfMergingPath configuration.
Made both references for new config and existing one lazy. It is 
definitely better not to block the whole system deployment for a 
feature that is not critical. Runtime feedback is acceptable.

Making Option fail with exception might seem against the very goal
of using options. In this specific case it is intended that dealing
with missing configuration exception is triggered in a single place.
If need arises it is very quickly retrofitted to simply returning 
the original Option[String] and manage None whereever used.